Add --stdin-file option
When passing code in via STDIN, Rubocop operates under the assumption
that the file belongs to the current working director, and looks for
.rubocop.yml files from there. This isn't always the case however, and
the --stdin-file flag allows you specify the real path to the file the
source belongs to.

This ensures that any .rubocop.yml files related to the original file
on disk is respected.